Building backend for employees and textiles, added feature for pictures but it's not perfect...

+++ src/main/java/fon/bg/ac/rs/retailApp/models/Textile.java
@@ -27,7 +27,7 @@ public class Textile {
     @Column(name = "id")
     private Integer id;
     private String uniqueCode;
-    private double piecePrice;
+    private int piecePrice;
     private String specialDescription;
     private int availableQuantity;
     private String photo;
